More code review fixes
