+++ dali/imgcodec/CMakeLists.txt
@@ -2,7 +2,7 @@ project(dali_imgcodec CUDA CXX C)
 
 set(CMAKE_CUDA_SEPARABLE_COMPILATION ON)
 
-add_subdirectory(formats)
+add_subdirectory(parsers)
 
 # Get all the source files and dump test files
 collect_headers(DALI_INST_HDRS PARENT_SCOPE)
