+++ dali/imgcodec/parsers/bmp.cc
@@ -1,4 +1,3 @@
-
 // Copyright (c) 2022, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,21 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef DALI_IMGCODEC_FORMATS_BMP_H_
-#define DALI_IMGCODEC_FORMATS_BMP_H_
-
-#include "dali/imgcodec/image_format.h"
+#include "dali/imgcodec/parsers/bmp.h"
 
 namespace dali {
 namespace imgcodec {
 
-class DLL_PUBLIC BmpParser : public ImageParser {
- public:
-  ImageInfo GetInfo(ImageSource *encoded) const override;
-  bool CanParse(ImageSource *encoded) const override;
-};
+ImageInfo BmpParser::GetInfo(ImageSource *encoded) const {
+  ImageInfo info;
+  // info.shape = {h, w, c};   // TODO(janton)
+  return info;
+}
 
-#endif  // DALI_IMGCODEC_FORMATS_BMP_H_
+bool BmpParser::CanParse(ImageSource *encoded) const {
+  return false;  // TODO(janton)
+}
 
 }  // namespace imgcodec
 }  // namespace dali
